Delete: hello.html
This was only a test file and wasn't really needed. So I have decided to
remove it entirely. I think this makes sense because it was totally
worthless keeping it.

If you disagree then I don't care because this is my website and I don't
care much for your oppinion if you disagree with what i do on my own
site. So there. Loser.